+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/bigquery/BigQueryQuerySourceDef.java
@@ -26,6 +26,7 @@ import com.google.api.services.bigquery.model.TableReference;
 import com.google.api.services.bigquery.model.TableSchema;
 import java.io.IOException;
 import java.io.ObjectInputStream;
+import java.util.Optional;
 import java.util.concurrent.atomic.AtomicReference;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.options.ValueProvider;
@@ -43,6 +44,7 @@ class BigQueryQuerySourceDef implements BigQuerySourceDef {
   private final Boolean useLegacySql;
   private final BigQueryIO.TypedRead.QueryPriority priority;
   private final String location;
+  private final String queryTempDataset;
   private final String kmsKey;
 
   private transient AtomicReference<JobStatistics> dryRunJobStats;
@@ -54,9 +56,17 @@ class BigQueryQuerySourceDef implements BigQuerySourceDef {
       Boolean useLegacySql,
       BigQueryIO.TypedRead.QueryPriority priority,
       String location,
+      String queryTempDataset,
       String kmsKey) {
     return new BigQueryQuerySourceDef(
-        bqServices, query, flattenResults, useLegacySql, priority, location, kmsKey);
+        bqServices,
+        query,
+        flattenResults,
+        useLegacySql,
+        priority,
+        location,
+        queryTempDataset,
+        kmsKey);
   }
 
   private BigQueryQuerySourceDef(
@@ -66,6 +76,7 @@ class BigQueryQuerySourceDef implements BigQuerySourceDef {
       Boolean useLegacySql,
       BigQueryIO.TypedRead.QueryPriority priority,
       String location,
+      String queryTempDataset,
       String kmsKey) {
     this.query = checkNotNull(query, "query");
     this.flattenResults = checkNotNull(flattenResults, "flattenResults");
@@ -73,6 +84,7 @@ class BigQueryQuerySourceDef implements BigQuerySourceDef {
     this.bqServices = bqServices;
     this.priority = priority;
     this.location = location;
+    this.queryTempDataset = queryTempDataset;
     this.kmsKey = kmsKey;
     dryRunJobStats = new AtomicReference<>();
   }
@@ -113,19 +125,26 @@ class BigQueryQuerySourceDef implements BigQuerySourceDef {
         useLegacySql,
         priority,
         location,
+        queryTempDataset,
         kmsKey);
   }
 
   void cleanupTempResource(BigQueryOptions bqOptions, String stepUuid) throws Exception {
+    Optional<String> queryTempDatasetOpt = Optional.ofNullable(queryTempDataset);
     TableReference tableToRemove =
         createTempTableReference(
-            bqOptions.getProject(), createJobIdToken(bqOptions.getJobName(), stepUuid));
+            bqOptions.getProject(),
+            createJobIdToken(bqOptions.getJobName(), stepUuid),
+            queryTempDatasetOpt);
 
     BigQueryServices.DatasetService tableService = bqServices.getDatasetService(bqOptions);
     LOG.info("Deleting temporary table with query results {}", tableToRemove);
     tableService.deleteTable(tableToRemove);
-    LOG.info("Deleting temporary dataset with query results {}", tableToRemove.getDatasetId());
-    tableService.deleteDataset(tableToRemove.getProjectId(), tableToRemove.getDatasetId());
+    if (queryTempDatasetOpt.isPresent()) {
+      // Remove temporary dataset only if it was created by Beam
+      LOG.info("Deleting temporary dataset with query results {}", tableToRemove.getDatasetId());
+      tableService.deleteDataset(tableToRemove.getProjectId(), tableToRemove.getDatasetId());
+    }
   }
 
   /** {@inheritDoc} */
